+++ R/fracreg.pe.table.R
@@ -3,7 +3,7 @@ fracreg.pe.table <- function(PE.p,PE.sd,PE.type,which.x,xvar.names,title,at)
 	z.ratio <- PE.p/PE.sd
 	p.value <- 2*(1-pnorm(abs(z.ratio)))
 
-	res <- cbind(Estimate = PE.p, `Std. Error` = PE.sd, `z value` = z.ratio, `Pr(>|z|)` = p.value)
+	res <- cbind(`dy/dx` = PE.p, `Std. Error` = PE.sd, `z value` = z.ratio, `Pr(>|z|)` = p.value)
 	rownames(res) <- which.x
 	rownames(res)[rownames(res) == "(Intercept)"] <- "(Intercept)"
 
@@ -16,6 +16,7 @@ fracreg.pe.table <- function(PE.p,PE.sd,PE.type,which.x,xvar.names,title,at)
 	cat(.fracreg.center(title), "\n")
 	cat(.fracreg.sep(), "\n")
 	
+	if(!is.na(PE.sd[1])) cat("\nNote: Standard errors computed using the Delta method\n")
 	stats::printCoefmat(res, P.values = TRUE, has.Pvalue = TRUE, digits = 4, signif.legend = TRUE)
 	
 	cat(.fracreg.sep(), "\n")